### 2022-10-12 #4833 @hjwilli
Hi @arfon, the source code is located in the `/app` directory.
https://github.com/carnival-data/carnival/tree/master/app

The license file and optional docker configuration are located at the root of the project.
