### 2022-10-20 1489a42
Fix and add DOI per JOSS review
