+++ paper.bib
@@ -19,7 +19,8 @@ publisher="Springer Berlin Heidelberg",
 address="Berlin, Heidelberg",
 pages="430--434",
 abstract="The Linked Data cloud contains large amounts of RDF data generated from databases.",
-isbn="978-3-662-46641-4"
+isbn="978-3-662-46641-4",
+doi = {10.1007/978-3-662-46641-4_40}
 }
 
 @article{HARRIS2019103208,
@@ -29,7 +30,7 @@ volume = {95},
 pages = {103208},
 year = {2019},
 issn = {1532-0464},
-doi = {https://doi.org/10.1016/j.jbi.2019.103208},
+doi = {10.1016/j.jbi.2019.103208},
 url = {https://www.sciencedirect.com/science/article/pii/S1532046419301261},
 author = {Paul A. Harris and Robert Taylor and Brenda L. Minor and Veida Elliott and Michelle Fernandez and Lindsay O'Neal and Laura McLeod and Giovanni Delacqua and Francesco Delacqua and Jacqueline Kirby and Stephany N. Duda},
 keywords = {Medical informatics, Electronic data capture, Clinical research, Translational research},
@@ -44,7 +45,7 @@ number = {2},
 pages = {377-381},
 year = {2009},
 issn = {1532-0464},
-doi = {https://doi.org/10.1016/j.jbi.2008.08.010},
+doi = {10.1016/j.jbi.2008.08.010},
 url = {https://www.sciencedirect.com/science/article/pii/S1532046408001226},
 author = {Paul A. Harris and Robert Taylor and Robert Thielke and Jonathon Payne and Nathaniel Gonzalez and Jose G. Conde},
 keywords = {Medical informatics, Electronic data capture, Clinical research, Translational research},
@@ -78,7 +79,7 @@ abstract = {Research electronic data capture (REDCap) is a novel workflow method
   year = {2020},
   note = {Articles initially published in Journal of Biomedical Informatics: X 5-8, 2020},
   issn = {1532-0464},
-  doi = {https://doi.org/10.1016/j.yjbinx.2020.100086},
+  doi = {10.1016/j.yjbinx.2020.100086},
   url = {https://www.sciencedirect.com/science/article/pii/S2590177X20300214},
   author = {Hayden G. Freedman and Heather Williams and Mark A. Miller and David Birtwell and Danielle L. Mowery and Christian J. Stoeckert},
   keywords = {Clinical data, Biomedical ontologies, Common data model, Data interoperability, Semantic Web technologies},
