### 2022-11-03 #4833 @hjwilli
Hi JOSS team, my team came up with some [potential reviewers](https://docs.google.com/spreadsheets/d/1zMc-QG1l2hoDVCrbjqn4ezAAIlGgOihVGwU704qhX70/edit?usp=sharing) for this paper. Please let us know if there is anything you need from us to move things forward! Thank you!

mgiardino
chaoran-chen
KonradHoeffner
ThomasThelen
reality
kinow
wetneb
UTH-Tuan
