+++ docs/production-builds.md
@@ -1,27 +1,63 @@
 # Production Builds
 
+## Publishing to Maven Local
+
+Carnival artifacts can be published to your local Maven repository usually found on your file system in the directory `~/.m2` via the following command. 
+
+```Shell
+./gradlew :carnival-gradle:publishToMavenLocal 
+```
+
 ## Publishing to Maven
 
-* Maven Repository Manager: https://s01.oss.sonatype.org/#welcome
-* Maven: https://search.maven.org/artifact/io.github.carnival-data/carnival-core
+Carnival artifacts are published to [Maven](https://search.maven.org/search?q=io.github.carnival-data) via the [Nexus Repository Manager](https://s01.oss.sonatype.org).
 
 ### Configuration
-Copy `.env-template` to `.env` and update the file with your maven central credentials and private key information. The signing file should be .gpg format, and SIGNING_KEY_ID is usually the last 8 digits of the fingerprint. More detail about the signing plugin available [here](https://docs.gradle.org/7.4.1/userguide/signing_plugin.html#sec:signatory_credentials).
+Copy `.env-template` to `.env` and update the file with your maven central credentials and private key information. The signing file should be **.gpg** (not **.asc**) format.  `SIGNING_KEY_ID` is usually the last 8 digits of the fingerprint. More detail about the signing plugin available [here](https://docs.gradle.org/7.4.1/userguide/signing_plugin.html#sec:signatory_credentials).
 
 
 ### Publish to Snapshot Repository
-When the carnivalVersion specified in `app\gradle.properties` ends with "-SNAPSHOT", the package will be published to the snapshot repository. Previous releases with the same version can be overwritten.
-```
+When the carnivalVersion specified in `app/gradle.properties` ends with "-SNAPSHOT", the package will be published to the snapshot repository. Previous releases with the same version can be overwritten.
+
+#### Docker
+
+```Shell
 docker-compose -f docker-compose-publish-maven.yml up
+```
+
+#### Gradle
+
+From the `app/` directory:
+
+```Shell
+source ../.env
+./gradlew :carnival-gradle:publishAllPublicationsToCentralRepository 
 ```
 
 ### Publish Release Versions
 If the version number does not end with "-SNAPSHOT", the package will be published to the staging repository and must be manually approved.
 
-1. Run the following to publish to the staging repository:
+#### Docker
+
+Run the following to publish to the staging repository:
 ```
 docker-compose -f docker-compose-publish-maven.yml up
 ```
+
+#### Gradle
+
+From the `app/` directory:
+
+
+```Shell
+source ../.env
+```
+
+```Shell
+./gradlew publishAllPublicationsToCentralRepository -Psigning.secretKeyRingFile=${SIGNING_PRIVATE_DIR}/${SIGNING_PRIVATE_FILE} -Psigning.password=${SIGNING_PRIVATE_KEY_PASSWORD} -Psigning.keyId=${SIGNING_KEY_ID} -Pcentral.user=${CENTRAL_USER} -Pcentral.password=${CENTRAL_PASSWORD} --no-daemon --console=plain
+```
+
+#### Publish The Release
 
 1. Log into the [Maven Nexus Repository Manager](https://s01.oss.sonatype.org/#welcome)
 
@@ -32,7 +68,8 @@ docker-compose -f docker-compose-publish-maven.yml up
 1. If the validation is successfull, click "Release" to publish the release.
 
 
-### Publishing Production Builds to Github (Deprecated)
+
+## Publishing Production Builds to Github (Deprecated)
 
[…]
