### 2022-11-17 b209649
clean up the build process. remove maven publishing from gradle plugin. update the docs.
