### 2022-11-18 376d1ed
Clean up API for adding models to a Carnival. Documentation.
