+++ docs/index.md
@@ -3,29 +3,33 @@
 
 ## Contents
 
-### Introduction
-- [Summary](#summary)
-- [Statement of Need](#statement-need)
-- [Features](#features)
+- Introduction
+  - [Statement of Need](#statement-need)
+  - [Overview](#summary)
+  - [Features](#features)
+- Development
+  - [Carnival Library Development](#core-development)
+- Applications
+  - [Using the Carnival Library](#using-carnival)
+  - [Groovy Scripts](#script-development)
+  - [Micronaut Applications](#app-development)
+  - [Reference Documentation](#reference-docs)
+  - [Application Programmer Interface (API)](#api)
 
-### Development
-- [Core Development](#core-development)
-- [Application Development](#app-development)
-- [Application Programmer Interface](#api)
+## <a name="statement-need"></a>Statement of need
 
-## <a name="summary"></a>Summary
 Loading, cleansing, and organizing data can dominate the time spent on a data science project [[Forbes]](https://www.forbes.com/sites/gilpress/2016/03/23/data-preparation-most-time-consuming-least-enjoyable-data-science-task-survey-says/?sh=5eebf58e6f63)[[Anaconda]](https://www.anaconda.com/state-of-data-science-2020?utm_medium=press&utm_source=anaconda&utm_campaign=sods-2020&utm_content=report).  This phenomenon is exacerbated in human subjects research at an academic medical institution where data are very complex, reside in disparate repositories with varying levels of accesibility, are coded by separate yet overlapping coding systems, frequently rely on manual data entry, and change over time.  Data provenance and reproducibility of results are important factors in human subjects research.  It is no easy task to implement a robust consistent data pipeline with clear data provenance that can be rerun when source data change.  While there are several mature libraries and toolkits that enable visualization and statistical computation once the analytical data set is generated, there are comparatively fewer informatics offerings to address these concerns.  
 
 Existing ETL technologies such as [Microsoft SQL Server Integration Services](https://docs.microsoft.com/en-us/sql/integration-services/sql-server-integration-services?view=sql-server-ver15) help with data staging.  Similarly, data manipulation tools like [pandas](https://pandas.pydata.org) facilitate transformation of series and matrix data.  **Carnival** distinguishes itself by offering a lightweight data caching mechanism coupled with data manipulation services built on a property graph rather than arrays and data frames.  This unique combination empowers informatics programmers to build pipelines, utilities, and applications that are comparatively richer in semantics and provenance.
 
 Knowledge bases in Resource Description Framework (RDF) triplestores can be valuable tools to harmonize and enrich complex data.  While there are relational to RDF mappers, transforming source relational data into RDF triples is challenging.  Property graphs offer a middle ground between relational and RDF.  They lack the native ability to benefit from ontologies represented in RDF, but are more friendly to algorithmic computation.  
 
-**Carnival** was developed to provide programmers with tools to go from source data to a well formatted property graph and onward to RDF with rich semantics and provenance.  There are two main components to Carnival.  The first is a data caching mechanism that supports the efficient aggregation of data from disparate sources.  The second is a layer built on top of [Apache Tinkerpop](https://tinkerpop.apache.org) that seeks to provide more standardized and semantically driven methods of interacting with a property graph.
 
+## <a name="summary"></a>Overview
 
-## <a name="statement-need"></a>Statement of need
+**Carnival** is a semantically driven informatics toolkit that enables the aggregation of data from disparate sources into a unified property graph and provides mechanisms to model and interact with property graph data in well-defined ways.
 
[…]
