+++ app/carnival-core/README.md
@@ -1,0 +1,3 @@
+# Carnival Core
+
+Top level Carnival library that includes graph modeling, graph methods, and Carnival objects.
