### 2022-11-22 #4833 @gkthiruvathukal
@arfon Is there some editorialbot magic to re-run the gitinspector statistics? We have a proper repository now.
