### 2022-11-28 b7251af
cleanup carnival initialization methods, add addModel() layer, self contained tutorial
