+++ joss/paper.bib
@@ -1,0 +1,87 @@
+@InProceedings{10.1007/978-3-662-46641-4_40,
+author="Gupta, Shubham
+and Szekely, Pedro
+and Knoblock, Craig A.
+and Goel, Aman
+and Taheriyan, Mohsen
+and Muslea, Maria",
+editor="Simperl, Elena
+and Norton, Barry
+and Mladenic, Dunja
+and Della Valle, Emanuele
+and Fundulaki, Irini
+and Passant, Alexandre
+and Troncy, Rapha{\"e}l",
+title="Karma: A System for Mapping Structured Sources into the Semantic Web",
+booktitle="The Semantic Web: ESWC 2012 Satellite Events",
+year="2015",
+publisher="Springer Berlin Heidelberg",
+address="Berlin, Heidelberg",
+pages="430--434",
+abstract="The Linked Data cloud contains large amounts of RDF data generated from databases.",
+isbn="978-3-662-46641-4",
+doi = {10.1007/978-3-662-46641-4_40}
+}
+
+@article{HARRIS2019103208,
+title = {The REDCap consortium: Building an international community of software platform partners},
+journal = {Journal of Biomedical Informatics},
+volume = {95},
+pages = {103208},
+year = {2019},
+issn = {1532-0464},
+doi = {10.1016/j.jbi.2019.103208},
+url = {https://www.sciencedirect.com/science/article/pii/S1532046419301261},
+author = {Paul A. Harris and Robert Taylor and Brenda L. Minor and Veida Elliott and Michelle Fernandez and Lindsay O'Neal and Laura McLeod and Giovanni Delacqua and Francesco Delacqua and Jacqueline Kirby and Stephany N. Duda},
+keywords = {Medical informatics, Electronic data capture, Clinical research, Translational research},
+abstract = {The Research Electronic Data Capture (REDCap) data management platform was developed in 2004 to address an institutional need at Vanderbilt University, then shared with a limited number of adopting sites beginning in 2006. Given bi-directional benefit in early sharing experiments, we created a broader consortium sharing and support model for any academic, non-profit, or government partner wishing to adopt the software. Our sharing framework and consortium-based support model have evolved over time along with the size of the consortium (currently more than 3200 REDCap partners across 128 countries). While the “REDCap Consortium” model represents only one example of how to build and disseminate a software platform, lessons learned from our approach may assist other research institutions seeking to build and disseminate innovative technologies.}
+}
+
+@article{HARRIS2009377,
+title = {Research electronic data capture (REDCap)—A metadata-driven methodology and workflow process for providing translational research informatics support},
+journal = {Journal of Biomedical Informatics},
+volume = {42},
+number = {2},
+pages = {377-381},
+year = {2009},
+issn = {1532-0464},
+doi = {10.1016/j.jbi.2008.08.010},
+url = {https://www.sciencedirect.com/science/article/pii/S1532046408001226},
+author = {Paul A. Harris and Robert Taylor and Robert Thielke and Jonathon Payne and Nathaniel Gonzalez and Jose G. Conde},
+keywords = {Medical informatics, Electronic data capture, Clinical research, Translational research},
+abstract = {Research electronic data capture (REDCap) is a novel workflow methodology and software solution designed for rapid development and deployment of electronic data capture tools to support clinical and translational research. We present: (1) a brief description of the REDCap metadata-driven software toolset; (2) detail concerning the capture and use of study-related metadata from scientific research teams; (3) measures of impact for REDCap; (4) details concerning a consortium network of domestic and international institutions collaborating on the project; and (5) strengths and limitations of the REDCap system. REDCap is currently supporting 286 translational research projects in a growing collaborative network including 27 active partner institutions.}
+}
+
+@article{forbes1,
[…]
