+++ joss/paper.md
@@ -1,0 +1,69 @@
+---
+title: 'Carnival: JVM Property graph data unification toolkit'
+tags:
+- Groovy
+- Tinkerpop
+- property graph
+- graph model
+- etl
+- data cache
+authors: 
+  - name: David Birtwell
+    orcid: 0000-0002-3766-7568
+    affiliation: 2
+  - name: Heather Williams
+    orcid: 0000-0002-5079-1328
+    affiliation: 1
+  - name: Tom Hutchinson
+    orcid: 0000-0002-1002-0449
+    affiliation: 1
+  - name: Louis Lee
+    orcid: 0000-0002-7297-3915
+    affiliation: 2
+  - name: Christian Stoeckert
+    orcid: 0000-0002-5714-991X
+    affiliation: 1
+affiliations:
+ - name: University of Pennsylvania, Institute for Biomedical Informatics
+   index: 1
+ - name: USC Norris Cancer Center, USA
+   index: 2
+date: 12 May 2022
+bibliography: paper.bib
+---
+
+# Summary
+Research activities in data rich areas such as biomedical informatics face many data challenges including harmonizing complex and disparate data, integrating existing knowledge bases into data sets for manual or machine learning analysis, and reproducibility of results. Graphs are a powerful data structure for naturally describing complex data. Information about data providence can be embedded in the graph itself to aid in quality control and reproducibility. **Carnival** is a semantically driven informatics toolkit that enables the aggregation of data from disparate sources into a unified property graph and provides mechanisms to model and interact with the graph in well-defined ways inspired by the Open Biological and Biomedical Ontology (OBO) Foundry ontologies.  
+
+# Statement of need
+Loading, cleansing, and organizing data can dominate the time spent on a data science project Forbes[@forbes1]. This phenomenon is exacerbated in human subjects research at an academic medical institution where data are very complex, reside in disparate repositories with varying levels of accessibility, are coded by separate yet overlapping coding systems, frequently rely on manual data entry, and change over time. Data provenance and reproducibility of results are important factors in human subjects research. It is no easy task to implement a robust consistent data pipeline with clear data provenance that can be rerun when source data change. While there are several mature libraries and toolkits that enable visualization and statistical computation once the analytical data set is generated, there are comparatively fewer informatics offerings to address these concerns.
+
+Existing ETL technologies such as [Microsoft SQL Server Integration Services](https://docs.microsoft.com/en-us/sql/integration-services/sql-server-integration-services) help with data staging. Similarly, data manipulation tools like [pandas](https://pandas.pydata.org) facilitate transformation of series and matrix data. **Carnival** distinguishes itself by offering a lightweight data caching mechanism coupled with data manipulation services built on a property graph rather than arrays and data frames. This combination empowers informatics programmers to build pipelines, utilities, and applications that are comparatively richer in semantics and provenance.
+
+Knowledge bases in Resource Description Framework (RDF) triplestores can be valuable tools to harmonize and enrich complex data. Transforming source relational data into RDF triples reflecting a data model is challenging. While there exist relational-to-RDF mappers such as Karma[@10.1007/978-3-662-46641-4_40], the configuration process is labor intensive and the resulting triples may not match a data model particularly one of sufficient complexity.
+
+**Carnival** was developed to create domain-specific property graph data models, and provide tools to create robust pipelines to import and manage data in that model. There are two main components to Carnival. The primary component is a layer built on top of [Apache Tinkerpop](https://tinkerpop.apache.org) that seeks to provide more standardized and semantically driven methods of interacting with a property graph. An additional component is a data caching mechanism that supports the efficient aggregation of data from disparate sources.
+
+
+### Key Features
+- A graph modeling framework that ensures graph data remain consistent
+- A data caching mechanism to ease the computational burden of data aggregation during the development process and promotes data provenance
[…]
