### 2022-11-29 #4833 @augustearth
@arfon, I created a new branch [joss2](https://github.com/carnival-data/carnival/tree/joss2), which contains all the files in master, the paper files in the joss/ directory, and a working draft-pdf Github action.  Due to the way that the original joss branch was created, this was the easiest way to go.  Is it possible to switch to this new branch to complete the submission?
