### 2023-01-29 #4833 @majensen
Hi @KonradHoeffner -- would you have time and willingness to review [this work](#4833) for [JOSS](https://joss.theoj.org)? I would really appreciate your expertise - vielen Dank!
--@majensen
