### 2023-01-29 #4833 @kinow
Oi @majensen :wave:  !

>would you be able to review [this work](https://github.com/openjournals/joss-reviews/issues/4833#4833) for [JOSS](https://joss.theoj.org/)?

Sure thing, I had a quick read at the readme and it looks interesting to read & review :+1:

Cheers
