### 2023-01-30 #5117 @editorialbot
[REVIEW]: Carnival: JVM Property graph data unification toolkit
**Submitting author:** @hjwilli (<a href="http://orcid.org/0000-0002-5079-1328">Heather Williams</a>)
**Repository:** https://github.com/carnival-data/carnival
**Branch with paper.md** (empty if default branch): joss2
**Version:** 2.1.6
