### 2023-01-31 carnival-data/carnival#105 @kinow
Review JOSS paper text
Hi @hjwilli :wave: !

I will add my review of the [Carnival JOSS paper](https://github.com/openjournals/joss-reviews/issues/5117) text here :+1:

- [x] in the summary, on this part "Information about data providence", did you mean "data provenance"?
- [x] and in "Data pertaining to the set of study subjects was striped", s/striped/stripped? (or maybe trimmed?)
- [x] in "Uses", use cases are described in different scenarios; are these cases documented somewhere; in one it says the "The source data repositories included a relational data warehouse "... is that a GitHub source repository somewhere?
- [x] #111

Cheers
-Bruno
