### 2023-01-31 carnival-data/carnival#106 @kinow
Add CONTRIBUTING.md
Hi,

At the end of the `README` file there's some information for contributors, but perhaps this could be expanded to provide users with a clear and simple way to contribute. It should include what contributions are welcome (as is in the `README` file), how to build and what to build Carnival, if tests are expected, what type of test, any community guidelines (CLA, policies, etc), and anything else that might be helpful to contributors.

This is related to the [JOSS paper review](https://github.com/openjournals/joss-reviews/issues/5117) item:

>Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
