### 2023-02-01 carnival-data/carnival#107 @kinow
Make installation instructions more prominent in the documentation for users
Hi,

For the [JOSS paper review](https://github.com/openjournals/joss-reviews/issues/5117), I looked for the installation instructions of Carnival.

Initially it wasn't clear whether it was published to Maven Central or not, whether I should user Docker, etc.

I think now I've found the docs to install it, under the [getting started](https://carnival-data.github.io/carnival/app-dev-getting-started.html) page.

- [x] #108
- [x] #109
- [x] #110
