### 2023-02-01 carnival-data/carnival#107 @kinow
This issue is related to the final pending items in my review

> Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
> Functionality: Have the functional claims of the software been confirmed?
> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
