### 2023-02-01 carnival-data/carnival#105 @kinow
The third item in the list above, about use cases and the data or sources for these cases, is related to the following pending item in my review for the JOSS paper:

> Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
