### 2023-02-01 #5117 @kinow
Hi @majensen , @hjwilli

I've partially finished my review. I created issues in the Carnival source code repository, linking to this issue. I marked the items I've completed in my review as done, in my checklist :point_up: All the pending items are being mentioned in the issues I created.

@hjwilli feel free to ping me on any of those issues to discuss or suggest some fixes :+1:

Thanks!
-Bruno
