### 2023-02-02 b4f7ec8
Correct two component issues from #105
