### 2023-02-03 carnival-data/carnival#113 @KonradHoeffner
please add missing API documentation items
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/5117).

At <https://carnival-data.github.io/carnival/groovydoc/index.html> there are still many undocumented items.
Please provide documentation for all items in the API docs.
