### 2023-02-03 #5117 @KonradHoeffner
The code has 13 commits by Hayden Freedman, who is not on the author list. On the other hand the author list includes Louis Lee, who does not seem to be a contributor to the codebase. Christian Stoeckert on the other hand only has a single commit where one word in the readme seems to have changed. Can you explain the choice of authors?
