### 2023-02-18 #5117 @majensen
@hjwilli could you comment on the concern raised by @KonradHoeffner at https://github.com/openjournals/joss-reviews/issues/5117#issuecomment-1415997159?
