### 2023-03-10 #5117 @hjwilli
Hi @majensen, thank you for your patience. My team is still in the process of making updates to the paper and codebase based on feedback and hopes to address most of the issues in the next few weeks.

For the author list, Christian Stoeckert is an ontology expert who has been key collaborator in the development of Carnival for the lifetime of the project and has directly contributed to the paper and should be included as an author. Louis Lee is a newer member of the team who has participated software design process for carnival and the paper and has make significant contributions to the demo application for this submission and should be included as an author. Hayden is no longer an active member of the group but has make significant contributions to the codebase and we are reaching out to them for inclusion.
