### 2023-03-13 #5117 @KonradHoeffner
@hjwilli: Yes, please definitely do include Hayden then, as required by the ethics guidelines:

> all significant contributors should be included in the author list
