### 2023-03-16 d5d5a50
Update paper.md
Closes #112
