### 2023-03-29 carnival-data/carnival#113 @augustearth
Done.  Updated documentation are available [online](https://carnival-data.github.io/carnival/groovydoc/index.html).
