### 2023-03-30 carnival-data/carnival#113 @KonradHoeffner
I mean there are still items that don't have any documentation at all. I have never used Groovy but isn't there some command that lists those items? For example in Rust there is the "missing docs" annotation, so I think this should be possible in other languages as well.
