### 2023-03-30 carnival-data/carnival#113 @augustearth
Hi @KonradHoeffner.  I am unaware of a way to tag missing docs in Groovy.  The tool used to build Groovy documentation is GroovyDoc, which (I assume) is a layer over JavaDoc.  I have gone through our source code and attempted to document all the important classes, fields, and methods.  The version of Gradle that is currently used for this project creates docs for protected and private elements.  Some of these may not have full documentation, but also would not be expected to be used in client code.

Again, if you let us know know which portions of the API are under-documented, we can attempt to address.  Thanks.
