+++ README.md
@@ -1,5 +1,5 @@
 [![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](https://github.com/pennbiobank/pennai/carnival-public/master/LICENSE)
-[![Carnival CI](https://github.com/pmbb-ibi/carnival/actions/workflows/carnival_ci.yml/badge.svg)](https://github.com/pmbb-ibi/carnival/actions/workflows/carnival_ci.yml)
+[![Carnival CI](https://github.com/carnival-data/carnival/actions/workflows/carnival_ci.yml/badge.svg)](https://github.com/carnival-data/carnival/actions/workflows/carnival_ci.yml)
 <a href='https://coveralls.io/github/pmbb-ibi/carnival?branch=master'><img src='https://coveralls.io/repos/github/pmbb-ibi/carnival/badge.svg?branch=master' alt='Coverage Status' /></a>
 [![Maven Central](https://img.shields.io/maven-central/v/io.github.carnival-data/carnival-core.svg?label=Maven%20Central)](https://search.maven.org/artifact/io.github.carnival-data/carnival-core)
 
@@ -29,7 +29,7 @@ Carnival has three principal components: a graph modeling architecture, a cachin
 
 Name | Description
 --- | ---
-[carnival-core](app/carnival-core/README.md) | Basic Carnival framework. Implements the basic Carnival framework classes (vines, reapers, reasonsers, etc). Defines the basic carnival graph schema (processes, databases). [Core model](https://github.com/pmbb-ibi/carnival/blob/master/app/carnival-core/src/main/groovy/carnival/core/graph/Core.groovy)
+[carnival-core](app/carnival-core/README.md) | Basic Carnival framework. Implements the basic Carnival framework classes (vines, reapers, reasonsers, etc). Defines the basic carnival graph schema (processes, databases). [Core model](https://github.com/carnival-data/carnival/blob/master/app/carnival-core/src/main/groovy/carnival/core/graph/Core.groovy)
 [carnival-vine](app/carnival-vine/README.md) | Mechanisms to faciliate the aggregation of data with data caching.  
 [carnival-graph](app/carnival-graph/README.md) | Framework for defining carnival graph schemas (vertex and edge definitions). Contains the basic vertex, edge, and property classes.
 [carnival-util](app/carnival-util/README.md) | Standalone package that contains utility and helper classes such as data tables, reports, and SQL utilties, which are primarily used for dealing with relational data.
