### 2023-04-28 carnival-data/carnival#113 @th5
Thank you, Konrad. We are almost finished updating the public API documentation.
