### 2023-04-28 carnival-data/carnival#113 @th5
To view all items in the API documentation:

https://github.com/carnival-data/carnival
=> click link, under External Resources

GroovyDoc API Documentation
https://carnival-data.github.io/carnival/groovydoc/index.html
=> click link, middle top of page

Index
https://carnival-data.github.io/carnival/groovydoc/index.html?index-all.html

Here is what I found:

- [ ] [apply(org.gradle.api.Project)](https://carnival-data.github.io/carnival/groovydoc/carnival/gradle/CarnivalApplicationPlugin.html#apply(org.gradle.api.Project)) - Method in [CarnivalApplicationPlugin](https://carnival-data.github.io/carnival/groovydoc/carnival/gradle/CarnivalApplicationPlugin.html)
- [ ] [AppUtil()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/AppUtil.html#AppUtil()) - Constructor in [AppUtil](https://carnival-data.github.io/carnival/groovydoc/carnival/util/AppUtil.html)
- [ ] [create(java.lang.String)](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FieldName.html#create(java.lang.String)) - Method in [FieldName](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FieldName.html)
- [ ] [CodeRefGroup()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/CodeRefGroup.html#CodeRefGroup()) - Constructor in [CodeRefGroup](https://carnival-data.github.io/carnival/groovydoc/carnival/util/CodeRefGroup.html)
- [ ] [conf](https://carnival-data.github.io/carnival/groovydoc/carnival/core/CarnivalNeo4jConfiguration.Gremlin.Neo4j.html#conf) - Property in [CarnivalNeo4jConfiguration.Gremlin.Neo4j](https://carnival-data.github.io/carnival/groovydoc/carnival/core/CarnivalNeo4jConfiguration.Gremlin.Neo4j.html)
- [ ] [DataTable.FieldNameExtensions()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/DataTable.FieldNameExtensions.html#DataTable.FieldNameExtensions()) - Constructor in [DataTable.FieldNameExtensions](https://carnival-data.github.io/carnival/groovydoc/carnival/util/DataTable.FieldNameExtensions.html)
- [ ] [DataTable.MetaData()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/DataTable.MetaData.html#DataTable.MetaData()) - Constructor in [DataTable.MetaData](https://carnival-data.github.io/carnival/groovydoc/carnival/util/DataTable.MetaData.html)
- [ ] [DataTableFiles()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/DataTableFiles.html#DataTableFiles()) - Constructor in [DataTableFiles](https://carnival-data.github.io/carnival/groovydoc/carnival/util/DataTableFiles.html)
- [ ] [DataTableVineMethod()](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/DataTableVineMethod.html#DataTableVineMethod()) - Constructor in [DataTableVineMethod](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/DataTableVineMethod.html)
- [ ] [DataTableVineMethodCall()](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/DataTableVineMethodCall.html#DataTableVineMethodCall()) - Constructor in [DataTableVineMethodCall](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/DataTableVineMethodCall.html)
- [ ] [DefaultElementConstraint()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/DefaultElementConstraint.html#DefaultElementConstraint()) - Constructor in [DefaultElementConstraint](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/DefaultElementConstraint.html)
- [ ] [DefaultGraphValidator()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/DefaultGraphValidator.html#DefaultGraphValidator()) - Constructor in [DefaultGraphValidator](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/DefaultGraphValidator.html)
- [ ] [Definition()](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/Definition.html#Definition()) - Constructor in [Definition](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/Definition.html)
- [ ] [directories](https://carnival-data.github.io/carnival/groovydoc/carnival/core/CarnivalNeo4jConfiguration.Gremlin.Neo4j.Conf.Dbms.html#directories) - Property in [CarnivalNeo4jConfiguration.Gremlin.Neo4j.Conf.Dbms](https://carnival-data.github.io/carnival/groovydoc/carnival/core/CarnivalNeo4jConfiguration.Gremlin.Neo4j.Conf.Dbms.html)
- [ ] [EdgeConstraint()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/EdgeConstraint.html#EdgeConstraint()) - Constructor in [EdgeConstraint](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/EdgeConstraint.html)
- [ ] [EdgeDomainException()](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/EdgeDomainException.html#EdgeDomainException()) - Constructor in [EdgeDomainException](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/EdgeDomainException.html)
- [ ] [ExcelUtil.ExcelRow()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/ExcelUtil.ExcelRow.html#ExcelUtil.ExcelRow()) - Constructor in [ExcelUtil.ExcelRow](https://carnival-data.github.io/carnival/groovydoc/carnival/util/ExcelUtil.ExcelRow.html)
- [ ] [ExcelUtil()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/ExcelUtil.html#ExcelUtil()) - Constructor in [ExcelUtil](https://carnival-data.github.io/carnival/groovydoc/carnival/util/ExcelUtil.html)
- [ ] [extractCompoundMultiFeatures(java.util.Map)](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FeatureReport.html#extractCompoundMultiFeatures(java.util.Map)) - Method in [FeatureReport](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FeatureReport.html)
- [ ] [extractMultiFeatures(java.util.Map)](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FeatureReport.html#extractMultiFeatures(java.util.Map)) - Method in [FeatureReport](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FeatureReport.html)
- [ ] [FieldNameExtensions()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FieldNameExtensions.html#FieldNameExtensions()) - Constructor in [FieldNameExtensions](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FieldNameExtensions.html)
- [ ] [FilesUtil.1](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FilesUtil.1.html) - Class in [carnival.util](https://carnival-data.github.io/carnival/groovydoc/carnival/util/package-summary.html)
- [ ] [FilesUtil.1()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FilesUtil.1.html#FilesUtil.1()) - Constructor in [FilesUtil.1](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FilesUtil.1.html)
- [ ] [FilesUtil()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FilesUtil.html#FilesUtil()) - Constructor in [FilesUtil](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FilesUtil.html)
- [ ] [findAllSteps(Set<FeatureSetRecipeStep>, carnival.util.FeatureSetRecipeStep)](https://carnival-data.github.io/carnival/groovydoc/carnival/util/DataSetDescriptor.html#findAllSteps(Set%3CFeatureSetRecipeStep%3E,%20carnival.util.FeatureSetRecipeStep)) - Method in [DataSetDescriptor](https://carnival-data.github.io/carnival/groovydoc/carnival/util/DataSetDescriptor.html)
- [ ] [findFieldName(java.lang.String, Map<String, String>)](https://carnival-data.github.io/carnival/groovydoc/carnival/util/MappedDataTable.html#findFieldName(java.lang.String,%20Map%3CString,%20String%3E)) - Method in [MappedDataTable](https://carnival-data.github.io/carnival/groovydoc/carnival/util/MappedDataTable.html)
- [ ] [GenericDataTableVineMethod()](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/GenericDataTableVineMethod.html#GenericDataTableVineMethod()) - Constructor in [GenericDataTableVineMethod](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/GenericDataTableVineMethod.html)
- [ ] [getEdgeConstraints()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/DefaultGraphSchema.html#getEdgeConstraints()) - Method in [DefaultGraphSchema](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/DefaultGraphSchema.html)
- [ ] [getFieldName()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FieldName.html#getFieldName()) - Method in [FieldName](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FieldName.html)
- [ ] [GraphMethod()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GraphMethod.html#GraphMethod()) - Constructor in [GraphMethod](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GraphMethod.html)
- [ ] [GraphMethodCall()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GraphMethodCall.html#GraphMethodCall()) - Constructor in [GraphMethodCall](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GraphMethodCall.html)
- [ ] [GraphMethodDynamic()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GraphMethodDynamic.html#GraphMethodDynamic()) - Constructor in [GraphMethodDynamic](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GraphMethodDynamic.html)
- [ ] [GraphMethodList()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GraphMethodList.html#GraphMethodList()) - Constructor in [GraphMethodList](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GraphMethodList.html)
- [ ] [GraphMethodProcess()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GraphMethodProcess.html#GraphMethodProcess()) - Constructor in [GraphMethodProcess](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GraphMethodProcess.html)
- [ ] [GraphValidationError()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GraphValidationError.html#GraphValidationError()) - Constructor in [GraphValidationError](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GraphValidationError.html)
- [ ] [gremlin](https://carnival-data.github.io/carnival/groovydoc/carnival/core/CarnivalNeo4jConfiguration.html#gremlin) - Property in [CarnivalNeo4jConfiguration](https://carnival-data.github.io/carnival/groovydoc/carnival/core/CarnivalNeo4jConfiguration.html)
- [ ] [GremlinTraitUtilities()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GremlinTraitUtilities.html#GremlinTraitUtilities()) - Constructor in [GremlinTraitUtilities](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/GremlinTraitUtilities.html)
- [ ] [has(org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.__, carnival.graph.PropertyDefinition, java.lang.Object)](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/ext/TinkerpopAnonTraversalExtension.html#has(org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.__,%20carnival.graph.PropertyDefinition,%20java.lang.Object)) - Method in [TinkerpopAnonTraversalExtension](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/ext/TinkerpopAnonTraversalExtension.html)
- [ ] [Identifier()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/Identifier.html#Identifier()) - Constructor in [Identifier](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/Identifier.html)
- [ ] [in(org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.__, carnival.graph.EdgeDefinition)](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/ext/TinkerpopAnonTraversalExtension.html#in(org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.__,%20carnival.graph.EdgeDefinition)) - Method in [TinkerpopAnonTraversalExtension](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/ext/TinkerpopAnonTraversalExtension.html)
- [ ] [isa(org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.__, carnival.graph.EdgeDefinition)](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/ext/TinkerpopAnonTraversalExtension.html#isa(org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.__,%20carnival.graph.EdgeDefinition)) - Method in [TinkerpopAnonTraversalExtension](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/ext/TinkerpopAnonTraversalExtension.html)
- [ ] [IterativeCsvWriter()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/IterativeCsvWriter.html#IterativeCsvWriter()) - Constructor in [IterativeCsvWriter](https://carnival-data.github.io/carnival/groovydoc/carnival/util/IterativeCsvWriter.html)
- [ ] [JsonVineMethod()](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/JsonVineMethod.html#JsonVineMethod()) - Constructor in [JsonVineMethod](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/JsonVineMethod.html)
- [ ] [JsonVineMethodCall.Meta()](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/JsonVineMethodCall.Meta.html#JsonVineMethodCall.Meta()) - Constructor in [JsonVineMethodCall.Meta](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/JsonVineMethodCall.Meta.html)
- [ ] [JsonVineMethodCall.Result()](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/JsonVineMethodCall.Result.html#JsonVineMethodCall.Result()) - Constructor in [JsonVineMethodCall.Result](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/JsonVineMethodCall.Result.html)
- [ ] [JsonVineMethodCall()](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/JsonVineMethodCall.html#JsonVineMethodCall()) - Constructor in [JsonVineMethodCall](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/JsonVineMethodCall.html)
- [ ] [JsonVineMethodListCall()](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/JsonVineMethodListCall.html#JsonVineMethodListCall()) - Constructor in [JsonVineMethodListCall](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/JsonVineMethodListCall.html)
- [ ] [LegacyValidator()](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/LegacyValidator.html#LegacyValidator()) - Constructor in [LegacyValidator](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/LegacyValidator.html)
- [ ] [loadDataFromFile(java.io.File, carnival.util.DataTable)](https://carnival-data.github.io/carnival/groovydoc/carnival/util/DataTable.html#loadDataFromFile(java.io.File,%20carnival.util.DataTable)) - Method in [DataTable](https://carnival-data.github.io/carnival/groovydoc/carnival/util/DataTable.html)
- [ ] [loadMetaDataFromFile(java.io.File)](https://carnival-data.github.io/carnival/groovydoc/carnival/util/DataTable.html#loadMetaDataFromFile(java.io.File)) - Method in [DataTable](https://carnival-data.github.io/carnival/groovydoc/carnival/util/DataTable.html)
- [ ] [Log()](https://carnival-data.github.io/carnival/groovydoc/carnival/util/Log.html#Log()) - Constructor in [Log](https://carnival-data.github.io/carnival/groovydoc/carnival/util/Log.html)
- [ ] [MappedDataTableVineMethod()](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/MappedDataTableVineMethod.html#MappedDataTableVineMethod()) - Constructor in [MappedDataTableVineMethod](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/MappedDataTableVineMethod.html)
- [ ] [MappedDataTableVineMethodCall()](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/MappedDataTableVineMethodCall.html#MappedDataTableVineMethodCall()) - Constructor in [MappedDataTableVineMethodCall](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/MappedDataTableVineMethodCall.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/Base.EX.html#MAX_VALUE) - Field in [Base.EX](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/Base.EX.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/Base.PX.html#MAX_VALUE) - Field in [Base.PX](https://carnival-data.github.io/carnival/groovydoc/carnival/graph/Base.PX.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/CacheMode.html#MAX_VALUE) - Field in [CacheMode](https://carnival-data.github.io/carnival/groovydoc/carnival/vine/CacheMode.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/util/CodeSystem.html#MAX_VALUE) - Field in [CodeSystem](https://carnival-data.github.io/carnival/groovydoc/carnival/util/CodeSystem.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/core/Core.EX.html#MAX_VALUE) - Field in [Core.EX](https://carnival-data.github.io/carnival/groovydoc/carnival/core/Core.EX.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/core/Core.PX.html#MAX_VALUE) - Field in [Core.PX](https://carnival-data.github.io/carnival/groovydoc/carnival/core/Core.PX.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/core/Core.VX.html#MAX_VALUE) - Field in [Core.VX](https://carnival-data.github.io/carnival/groovydoc/carnival/core/Core.VX.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/DataSetDescriptorGraph.EX.html#MAX_VALUE) - Field in [DataSetDescriptorGraph.EX](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/DataSetDescriptorGraph.EX.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/DataSetDescriptorGraph.VX.html#MAX_VALUE) - Field in [DataSetDescriptorGraph.VX](https://carnival-data.github.io/carnival/groovydoc/carnival/core/graph/DataSetDescriptorGraph.VX.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/util/ExcelUtil.EXCEL.html#MAX_VALUE) - Field in [ExcelUtil.EXCEL](https://carnival-data.github.io/carnival/groovydoc/carnival/util/ExcelUtil.EXCEL.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FeatureReport.DataMode.html#MAX_VALUE) - Field in [FeatureReport.DataMode](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FeatureReport.DataMode.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FeatureReport.DateWords.html#MAX_VALUE) - Field in [FeatureReport.DateWords](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FeatureReport.DateWords.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FirstOrLastDateComparator.html#MAX_VALUE) - Field in [FirstOrLastDateComparator](https://carnival-data.github.io/carnival/groovydoc/carnival/util/FirstOrLastDateComparator.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/util/SingleValueDateComparator.html#MAX_VALUE) - Field in [SingleValueDateComparator](https://carnival-data.github.io/carnival/groovydoc/carnival/util/SingleValueDateComparator.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/carnival/util/SummaryDateComparator.html#MAX_VALUE) - Field in [SummaryDateComparator](https://carnival-data.github.io/carnival/groovydoc/carnival/util/SummaryDateComparator.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/test/TestModel.EX.html#MAX_VALUE) - Field in [TestModel.EX](https://carnival-data.github.io/carnival/groovydoc/test/TestModel.EX.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/test/TestModel.PX.html#MAX_VALUE) - Field in [TestModel.PX](https://carnival-data.github.io/carnival/groovydoc/test/TestModel.PX.html)
- [ ] [MAX_VALUE](https://carnival-data.github.io/carnival/groovydoc/test/TestModel.VX.html#MAX_VALUE) - Field in [TestModel.VX](https://carnival-data.github.io/carnival/groovydoc/test/TestModel.VX.html)
- [ ] [MetaData(carnival.util.GenericDataTable)](https://carnival-data.github.io/carnival/groovydoc/carnival/util/GenericDataTable.MetaData.html#MetaData(carnival.util.GenericDataTable)) - Constructor in [GenericDataTable.MetaData](https://carnival-data.github.io/carnival/groovydoc/carnival/util/GenericDataTable.MetaData.html)
[…]
