### 2023-05-09 ba08d79
Joss paper: fixed spelled
