### 2023-05-10 carnival-data/carnival#120 @KonradHoeffner
JOSS paper missing spaces
There are missing spaces in front of the left parenthese of citations, for example "project(Press, 2016)".

This issue is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/5117 and the closed issue https://github.com/carnival-data/carnival/issues/112 (that I cannot reopen).
