### 2023-05-11 29d6965
Update paper.md
Added Hayden Freedman to the authors list, updated wording, fixed spacing

References #120
