### 2023-05-11 a843547
Update paper.md
Whitespace updates

Ref #120
