+++ README.md
@@ -9,14 +9,10 @@
 
 ## External Resources
 
--   [Usage Website](https://carnival-data.github.io/carnival/)
--   [GroovyDoc API Documentation](https://carnival-data.github.io/carnival/groovydoc/index.html)
--   [Demonstration Project](https://github.com/carnival-data/carnival-micronaut)
+-   [Documentation](https://carnival-data.github.io/carnival/)
+-   [API Documentation](https://carnival-data.github.io/carnival/groovydoc/index.html)
+-   [Example Demonstration Project](https://github.com/carnival-data/carnival-micronaut)
 
-## Contents
-- [Overview](#overview)
-- [Packages](#packages)
-- [Contribution Guide](README.md#contribution-guide)
 
 ## <a name="overview"></a> Overview
 
@@ -29,20 +25,13 @@ Carnival has three principal components: a graph modeling architecture, a cachin
 
 Name | Description
 --- | ---
-[carnival-core](app/carnival-core/README.md) | Basic Carnival framework. Implements the basic Carnival framework classes (vines, reapers, reasonsers, etc). Defines the basic carnival graph schema (processes, databases). [Core model](https://github.com/carnival-data/carnival/blob/master/app/carnival-core/src/main/groovy/carnival/core/graph/Core.groovy)
-[carnival-vine](app/carnival-vine/README.md) | Mechanisms to faciliate the aggregation of data with data caching.  
+[carnival-core](app/carnival-core/README.md) | Basic Carnival framework. Implements the basic Carnival framework classes (vines, carnival modeling framework, carnival graph algorithm framework, etc). Defines the [core carnival graph model](https://github.com/carnival-data/carnival/blob/master/app/carnival-core/src/main/groovy/carnival/core/graph/Core.groovy). This model defines key carnival concepts such as processes, databases, and namespaces.
+[carnival-vine](app/carnival-vine/README.md) | Framework for data adaptors called vines which faciliate loading and aggregation of source data. Implements data caching facilities.
 [carnival-graph](app/carnival-graph/README.md) | Framework for defining carnival graph schemas (vertex and edge definitions). Contains the basic vertex, edge, and property classes.
 [carnival-util](app/carnival-util/README.md) | Standalone package that contains utility and helper classes such as data tables, reports, and SQL utilties, which are primarily used for dealing with relational data.
 [carnival-gradle](app/carnival-gradle/README.md) | Gradle plugin for building Carnival applications and libraries.
 
 ## <a name="contribution-guide"></a> Contribution Guide
-Carnival is an open source project and welcomes contributions! Here are some ways to help:
-
-* Writing and improving the documentation
-* Reporting or fixing bugs
-* Contributing features or enhancements
-* Creating carnival extensions for more domains
-* Creating demonstration projects or tutorials
-
+Carnival is an open source project and welcomes contributions! Please see the [Contribution Guide](CONTRIBUTING.md) for ways to contribute.
 
 
