### 2023-05-11 carnival-data/carnival#106 @hjwilli
Hi @kinow, CONTRIBUTIONS.md has been added, please let us know if this is sufficient.
