### 2023-05-11 carnival-data/carnival#106 @kinow
> Hi @kinow, CONTRIBUTIONS.md has been added, please let us know if this is sufficient.

Yes! Thank you @hjwilli ! Closing.
