### 2023-05-11 carnival-data/carnival#113 @augustearth
4959ad0e7368eff80b4ef80653a027b04e9bc7dc

- I have no idea where FilesUtil.1 is coming from.  It does not appear in the source.  Must be an artifact of compilation?
- all methods in TinkerpopAnonTraversalExtension have a @see tag
- valueOf(), values(), previous(), and next() are methods of all enums and should not need commenting
- there are @see references for CarnivalNeo4jConfiguration
