### 2023-05-12 e945249
docker-compose support for example scripts
Added docker-compose file that can be used to start an interactive shell w. groovy available on the command line, updated documentation.

Ref #110
