+++ index.md
@@ -285,6 +285,9 @@ The example scripts use [Groovy Grape](http://docs.groovy-lang.org/latest/html/d
 groovy -Dgroovy.grape.report.downloads=true graph-model-1.groovy
 ```
 
+> **Note**
+> To assist with running the example scripts, `docs/groovy` contains a docker-compose file to start an interactive shell with groovy available on the command line as an alternative to installing Groovy. To start the shell, from the example directory run `docker-compose run runner bash`. Scripts can be run with commands such as `groovy graph-model-1.groovy`.
+
 
 ### <a name="app-development"></a>Groovy Applications
 The Carnival library can be included in Gradle Groovy applications and libraries.  
