### 2023-05-12 c985c75
Update index.md
better wording
