### 2023-05-12 carnival-data/carnival#107 @hjwilli
Hi @kinow, I believe these issues have now all been addressed, and the documentation on how to use carnival in a project should be clearer:
https://carnival-data.github.io/carnival/#using-carnival

Please let us know if this is sufficient!
