### 2023-05-12 carnival-data/carnival#105 @hjwilli
Hi @kinow, regarding the third item, the data warehouse project is about human subject research and involves proprietary data schemas and protected information, and cannot be made public. The [demonstration project](https://github.com/carnival-data/carnival-demo-biomedical) referenced in the paper is public and is intended to help meet the reproducibility need, it contains synthetic clinical data in relational format.
