### 2023-05-14 carnival-data/carnival#107 @kinow
The installation/usage instructions look good to me. I've used Groovy+Grapes in the past, and I believe users and readers would be able to read Grapes' docs (if needed) and follow the steps in the new docs. Thanks @hjwilli ! Closing.
