### 2023-05-14 carnival-data/carnival#105 @kinow
> Hi @kinow, regarding the third item, the data warehouse project is about human subject research and involves proprietary data schemas and protected information, and cannot be made public. The [demonstration project](https://github.com/carnival-data/carnival-demo-biomedical) referenced in the paper is public and is intended to help meet the reproducibility need, it contains synthetic clinical data in relational format.

@hjwilli maybe the information above should be included in the paper? I looked at the currently rendered version in the JOSS paper pull request, and I can see the mention of the demonstration project, but you start saying "Additionally, the demonstration project...". IMHO, it would be best precede that with a note that the use cases mentioned have proprietary data, and that the demonstration project is public, etc. More or less what you said above.

WDYT?
