### 2023-05-15 carnival-data/carnival#113 @hjwilli
> *  I have no idea where FilesUtil.1 is coming from. It does not appear in the source. Must be an artifact of compilation?

It seems like this is from an [anonymous inner class](https://groovy-lang.org/objectorientation.html#_anonymous_inner_class) defined [here](https://github.com/carnival-data/carnival/blob/c985c7585601a2def6fac973c20fb4f50243169b/app/carnival-util/src/main/groovy/carnival/util/FilesUtil.groovy#L27), it should be safe to ignore.
