### 2023-05-22 carnival-data/carnival#122 @KonradHoeffner
difference in demo application links between paper and README.md
In the JOSS paper, the repository <https://github.com/carnival-data/carnival-demo-biomedical> is linked, however [README.md](https://github.com/carnival-data/carnival/blob/master/README.md) links to <https://github.com/carnival-data/carnival-demo-micronaut> which seems to be incomplete. For example I didn't find a Dockerfile or docker-compose.yml and the README.md there is very short.

Please make sure that there is a unified link to a working demo application.

This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/5117).
