### 2023-05-22 carnival-data/carnival#123 @KonradHoeffner
JOSS paper–functional claims
I have nearly finished the JOSS review but I have a hard time verifying the functional claims and have some suggestions for improvement:

## 1. Show more in the demo

The current demonstration using Docker Compose works well, but it lacks certain features
* The provided links only display a very small amount of JSON data (7 patients with 3 attributes each) without any interactive options or additional details.
* The link on http://localhost:5858/ mentions over 38000 different conditions but doesn't provide a way to view or export them. Consider adding a link or a method to access this information.
* JSON outputs lack hyperlinks, making it challenging to traverse the graph. Consider enhancing the JSON output with clickable links.

## 2. Expand on relationship with RDF triple stores and ontologies

Please elaborate on the relationship between your approach and RDF triple stores and ontologies. Does your software integrate with them or propose a separate alternative? Provide details on the integration or alternative approach in the paper or documentation.

## 3. Specify data export options

Please describe the available data export formats and methods. Can the resulting data be exported in RDF serializations like Turtle, N-Triples, or RDF/XML? Can it be exported as OLAP Cubes? Is there a custom JSON API? Does it utilize the existing API structure from Apache Tinkerpop? Explain these options and their availability in the demo's API.

## 4. Provide an introduction to Apache Tinkerpop
Briefly introduce Apache Tinkerpop and its relevance to your software. Consider referencing a suitable introductory paper about Apache Tinkerpop if one exists.

## 5. Explain data viewing and traversal capabilities
Describe how users can view and traverse the resulting data. Provide details on any available visualization tools, interactive interfaces, or navigation methods.

## 6. Highlight advantages over RDF-based approaches like Karma and Tarql
Clearly explain how your approach is less labor-intensive compared to existing tools like Karma and Tarql. Provide an exemplary use case where these tools would fail or be too labor-intensive while your approach succeeds.

## 7. Provide links and configuration details for analytical data sets
For the "Production of analytical data sets" use case, link to the relevant "analytical data sets for human subjects research" described in the paper, along with any related papers. Include the carnival configuration and input specifics for this use case.

## 8. General expansion

Just in general, I think the paper is too short to be easily understandable (at least from my point of view) and would profit from adding a lot more preliminaries, examples (or one running example) and additional scientific literature references.

This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/5117).
