### 2023-06-11 #5117 @kinow
Hi @majensen ,

I think I have one issue open pending response in their repository: https://github.com/carnival-data/carnival/issues/105

I'm waiting until the issues I created are resolved, and then I'll check the last box and finish my review (the one linked above is about use cases/docs/maybe a little repro, so I left that unchecked)
