### 2023-10-30 82b7c97
README: Updated core packages table, adding links to the relevant API docs and links to source code. Removed links to empty package readmes
