### 2024-06-28 d5ec8ea
Update paper.md
Updated text in the 'Use' section to note that the use cases described involve human subject data research and proprietary data and cannot be made public, so we provided a demonstration project to meet the reproducibility need.

See #105
