### 2024-06-28 ae9ed40
Update paper.md
In the description of the demonstration project, remove the text around reasoning and describe more specifically the carnival features that are being demonstrated.

Ref #111
