+++ joss/paper.md
@@ -69,6 +69,6 @@ We drew upon ontology modeling in the OBO Foundry as inspiration for the Carniva
 #### System integrations
 **Carnival's** ability to integrate data from disparate resources into a flexible computational resource enables data driven system integrations. For example, **Carnival** has been used effectively to integrate a custom help desk ticketing system with Monday.com. The help desk ticketing system was developed locally with a back-end relational database. Monday.com is accessible via its API for reads and writes. By modeling the help desk and Monday.com data as separate graph models, then using a third graph model to integrate the two, a **Carnival** integration application was developed to integrate the two data sets and compute changes that needed to occur based on the state of the data. In this example, **Carnival** was partnered with [Micronaut](https://micronaut.io) and deployed as a [Docker](https://www.docker.com) container on [Microsoft Azure](https://azure.microsoft.com/). The service would build an in-memory graph at regular minute intervals, compute the changes that were required, then call the appropriate web services to execute the logic of the integration.
 
-Because some of the use cases above involve human subject research and contain proprietary data they cannot be made public. Therefore, an example repository [carnival-demo-biomedical](https://github.com/carnival-data/carnival-demo-biomedical) has been provided that demonstrates the integration of Carnival with Micronaut and Docker. This example project models and integrates synthetic electronic health record data from a SQL database and CSV files, reasons over the data to create case and control patient cohorts, and presents an API endpoint to the user to explore the data.
+Because some of the use cases above involve human subject research and contain proprietary data they cannot be made public. Therefore, an example repository [carnival-demo-biomedical](https://github.com/carnival-data/carnival-demo-biomedical) has been provided that demonstrates the integration of Carnival with Micronaut and Docker. This example project models and integrates synthetic electronic health record data from a SQL database and CSV files, demonstrates the provenance tracking of the carnival graph algorithm framework with methods with that create case and control patient cohorts, and presents an API endpoint to the user to explore the data.
 
 # References
