### 2024-06-28 da6bd40
Update README.md
Fix obsolete demo project link

Ref #122
