### 2024-06-28 #5117 @hjwilli
Hi @crvernon, I've made changes to address the open issues raised by @kinow (https://github.com/carnival-data/carnival/issues/105), and one of the issues by @KonradHoeffner (https://github.com/carnival-data/carnival/issues/122). These should now be resolved. Working through addressing https://github.com/carnival-data/carnival/issues/123.
