### 2024-06-28 fdb01ad
Update paper.md
Fix typo

Ref #111
