### 2024-07-22 #5117 @majensen
@hjwilli maybe you can give an overview of your progress so far on carnival-data/carnival#123. We may be able to negotiate a path forward; I think we all would love to set this paper free.
