### 2024-07-25 #5117 @hjwilli
> @hjwilli maybe you can give an overview of your progress so far on [carnival-data/carnival#123](https://github.com/carnival-data/carnival/issues/123). We may be able to negotiate a path forward; I think we all would love to set this paper free.

Thank you, I am working on a partial response and updates related to this that will be pushed next week, and would appreciate feedback at that point!
