### 2024-08-03 #5117 @hjwilli
Hi @majensen, as an update, I've made extensive updates this week for https://github.com/carnival-data/carnival/issues/123, a few small things left.
