### 2024-08-07 d52575a
Update graph-api.md
Clarification on the relationship between the graph api, tinkerpop, and models. Links to tinkerpop documentation.

Ref #123
