### 2024-09-20 carnival-data/carnival#123 @hjwilli
Thanks for your feedback, I've made some changes that hopefully will help resolve this.

**1. Show more in the demo**
The api for the demo has been expanded to include endpoints for individual patient and encounter nodes (`/patient/{id}`, `/encounter/{id}`), the patient and encounter json now provide details on the associated meds and conditions, and URIs of connected nodes have been incorporated into the responses. Because JSON data is being returned it is browser dependent if they are clickable, but copy/paste would work.

Additionally, api endpoints have been added to export the graph in graphml or graphson format.

**2. Expand on relationship with RDF triple stores and ontologies**
I updated the demo documentation to elaborate on this, particularly in the ability to add a model to a property graph:
https://github.com/carnival-data/carnival-demo-biomedical/blob/master/docs/walkthrough.md#defining-the-graph-model

This is also discussed in the main carnival documentation here:
https://carnival-data.github.io/carnival/graph-model.html

**3. Specify data export options**
The graph can be exported using the Tinkerpop api. I implemented graph export options in the demo and updated the demo documentation:
https://github.com/carnival-data/carnival-demo-biomedical/blob/master/docs/walkthrough.md#exporting-the-graph-for-analysis

**4. Provide an introduction to Apache Tinkerpop**
I updated the main and demo documentation to talk more about tinkerpop and linked to introductory documentation:

https://github.com/carnival-data/carnival-demo-biomedical/blob/master/docs/walkthrough.md#carnival-developer-walkthrough

https://carnival-data.github.io/carnival/graph-api.html

**5. Explain data viewing and traversal capabilities**
Describe how users can view and traverse the resulting data. Provide details on any available visualization tools, interactive interfaces, or navigation methods.

I updated the demo api to have more traversal capability gave and example of how the api or graph export could be used for visualization in the demo documentation:
https://carnival-data.github.io/carnival/graph-api.html

**6. Highlight advantages over RDF-based approaches like Karma and Tarql**

Carnival is not meant to be a competitor to RDF graph tools, it is an attempt to bring some of the strengths of RDF to property graphs by incorporating semantic features inherent to RDF and OBO foundry ontologies in its modeling framework, and doing so in a way that can be incorporated into JVM based ETL pipelines.

Property and RDF graphs have different advantages and disadvantages. For example, the semantics of RDF graphs allow for greater standardization and interoperability and allow much easier harmonization of new data domains, however they require more specialized knowledge to use effectively, and can be less performant. Property graphs tend to be simpler, have a wider array of commercial options available, but do not have any inherent modeling or semantic capabilities.

The labor-intensity or feasibility of incorporating rdf vs property graphs would be very specific to the use case, involving things like the expertise of the developers and subject matter experts involved, organizational constraints on allowed technologies, existing infrastructure, frequency and size of graphs being built or re-built, what the intended use is of the graph data, etc.

**7. Provide links and configuration details for analytical data sets**
For the "Production of analytical data sets" use case, link to the relevant "analytical data sets for human subjects research" described in the paper, along with any related papers. Include the carnival configuration and input specifics for this use case.

Changes were made to address this as described here:
https://github.com/carnival-data/carnival/issues/105#issuecomment-1546936916

Unfortunately, these examples involve human subject research and involve proprietary schemas and protected health information and cannot be made public. The demo is intended to meet this need.

**8. General expansion**
I updated the demonstration project, documentation and paper to hopefully provide a clearer view of how this project can be used to create graph-based ETL processes to generate property graphs that incorporate a layer of semantic modeling.
