### 2024-09-21 #5117 @majensen
Thanks for this work @hjwilli. My opinion is that the additional documentation and example provided and described in [the response](https://github.com/carnival-data/carnival/issues/123#issuecomment-2364660204) to carnival-data/carnival#123 is very detailed and satisfies me. @KonradHoeffner, do you agree? (JOSS papers themselves are meant to be quite short, supplemented in documentation as in this instance.)
