### 2024-09-23 carnival-data/carnival#123 @KonradHoeffner
- [ ] 1. Show more in the demo
 - [ ] 2. Expand on relationship with RDF triple stores and ontologies
 - [ ] 3. Specify data export options
 - [ ] 4. Provide an introduction to Apache Tinkerpop
 - [ ] 5. Explain data viewing and traversal capabilities
 - [ ] 6. Highlight advantages over RDF-based approaches like Karma and Tarql
 - [ ] 7. Provide links and configuration details for analytical data sets
