### 2024-09-23 #5117 @KonradHoeffner
@majensen: Of the two leftover major issue, <https://github.com/carnival-data/carnival/issues/123> has been sufficiently addressed but <https://github.com/carnival-data/carnival/issues/113> is still open and there seem to remain undocumented items in the API.
