### 2024-09-27 #5117 @majensen
Thanks @KonradHoeffner - I've had a look at the documentation, my feeling is that major classes and "user facing" methods have a reasonable amount of explanation. The general documentation has also been expanded significantly. Just from my perspective, I think documentation also evolves and interaction with users will expose those components that are underdocumented. I'm inclined to recommend moving forward as-is, but it is up to you. If there are specific elements that you feel need docs, can you highlight these in the issue?
