### 2024-10-05 #5117 @majensen
@KonradHoeffner thanks for your comment. I do agree, and maybe the right direction would be to expose less. But I would like to move forward, since we are approaching 2 years in review, and I believe the functionality is established and the general documentation has been significantly improved. @hjwilli - I will review the paper itself and may have some minor updates to apply.
