### 2024-10-05 #5117 @majensen
The only minor proof comment I have on the paper is, can we make the 'Features' paragraph at l.50 into a sentence; keep the bullets but add "The main features of Carnival are:" before the list, and punctuate the bulleted sentences with comma, comma, period.
