### 2024-10-15 c32e20c
update docker-compose
remove deprecated 'version' keyword

closes #131
