### 2024-10-15 68344fe
Update paper.md
Minor update to wording of "key features" section
