### 2024-10-19 #5117 @majensen
Thanks a lot @hjwilli - for some reason, the bullets didn't render right in lines 49-53, maybe check that. [I would also put a comma between "model" and "particularly" at line 43.]
For publication, we need an open archive of the code repo, using Zenodo or similar. There is a quick tutorial [here](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729) if you need it. After you've made the repo, can you report the DOI and version number here?
