### 2024-11-27 cab5a63
Update paper.md
typo, spacing
