### 2024-11-27 a92c42f
Update carnival_ci.yml
