+++ .github/workflows/carnival_ci.yml
@@ -7,7 +7,7 @@ on:
 jobs:
   test_runner:
     name: Unit Tests
-    runs-on: ubuntu-18.04
+    runs-on: ubuntu-latest
 
     steps:
     - uses: actions/checkout@v2
