### 2024-11-27 112390b
Update carnival_ci.yml
