+++ .github/workflows/carnival_ci.yml
@@ -13,12 +13,12 @@ jobs:
     - uses: actions/checkout@v2
     
     - name: Build testing container
-      run: docker-compose -f ./docker-compose-test.yml build
+      run: docker compose -f ./docker-compose-test.yml build
     
     - name: Run tests and publish coverage results
 
       run: |
-        docker-compose -f docker-compose-test.yml run -e COVERALLS_REPO_TOKEN -w /appsrc/app app gradle test coveralls --console=plain --continue -Dtest.http=false
+        docker compose -f docker-compose-test.yml run -e COVERALLS_REPO_TOKEN -w /appsrc/app app gradle test coveralls --console=plain --continue -Dtest.http=false
       env:
        COVERALLS_REPO_TOKEN: ${{ secrets.COVERALLS_REPO_TOKEN }}
 
@@ -35,10 +35,10 @@ jobs:
 
       - name: Build Docker Images
         run: |
-          docker-compose -f ./docker-compose-doc.yml build
+          docker compose -f ./docker-compose-doc.yml build
       - name: Building Docs
         run: |
-          docker-compose -f ./docker-compose-doc.yml up --abort-on-container-exit --force-recreate
+          docker compose -f ./docker-compose-doc.yml up --abort-on-container-exit --force-recreate
           rm docs/.gitignore
       - name: Deploy
         uses: JamesIves/github-pages-deploy-action@3.7.1
