+++ README.md
@@ -1,4 +1,4 @@
-[![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](https://github.com/pennbiobank/pennai/carnival-public/master/LICENSE)
+[![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](https://github.com/carnival-data/carnival/blob/master/LICENSE)
 [![Carnival CI](https://github.com/carnival-data/carnival/actions/workflows/carnival_ci.yml/badge.svg)](https://github.com/carnival-data/carnival/actions/workflows/carnival_ci.yml)
 <a href='https://coveralls.io/github/pmbb-ibi/carnival?branch=master'><img src='https://coveralls.io/repos/github/pmbb-ibi/carnival/badge.svg?branch=master' alt='Coverage Status' /></a>
 [![Maven Central](https://img.shields.io/maven-central/v/io.github.carnival-data/carnival-core.svg?label=Maven%20Central)](https://search.maven.org/artifact/io.github.carnival-data/carnival-core)
