### 2025-04-13 #5117 @majensen
Thanks @hjwilli - to meet a JOSS requirement, would yo be able to change te archive name to the paper's title?
